# optimize performance of regex operations