--- a/pkg/api/error.go
+++ b/pkg/api/error.go
@@ -57,12 +57,13 @@ func errorMsg(message string, code int) *models.Error {
 	}
 }
 
+var re = regexp.MustCompile("^(.*)Params$")
+
 func handleRekorAPIError(params interface{}, code int, err error, message string, fields ...interface{}) middleware.Responder {
 	if message == "" {
 		message = http.StatusText(code)
 	}
 
-	re := regexp.MustCompile("^(.*)Params$")
 	typeStr := fmt.Sprintf("%T", params)
 	handler := re.FindStringSubmatch(typeStr)[1]