profiling Rekor yielded that we were doing a lot of regex allocations in the hot path; this change optimizes those down with pretty big results:

```sh
goos: darwin
goarch: arm64
pkg: github.com/sigstore/rekor/pkg/types/hashedrekord/v0.0.1
cpu: Apple M4 Pro
BenchmarkHashValidation/Old/SHA256-14             171613              5854 ns/op           18185 B/op        173 allocs/op
BenchmarkHashValidation/New/SHA256-14           36960652                31.70 ns/op           32 B/op          1 allocs/op
BenchmarkHashValidation/Old/SHA384-14             143301              9013 ns/op           29637 B/op        239 allocs/op
BenchmarkHashValidation/New/SHA384-14           29359563                41.33 ns/op           48 B/op          1 allocs/op
BenchmarkHashValidation/Old/SHA512-14             113379             10810 ns/op           34810 B/op        303 allocs/op
BenchmarkHashValidation/New/SHA512-14           22802001                51.42 ns/op           64 B/op          1 allocs/op
PASS
ok      github.com/sigstore/rekor/pkg/types/hashedrekord/v0.0.1 7.834s
```